+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Javis.jl - Changelog
 
-##for PR
-- change font scaling in svg2luxor from 1/2 to 425/1000
+## Unreleased
+- Shape.jl
+  - Now handles holes and polys both ways clockwise and counter-cw , this can happend if we are transforming that has an y-flip transform (typically like svg's that dvisvgm makes),
+- latex.jl
+  - two module variables `LaTeXusepackages` and `LaTeXprog` to specify LaTeX behavior
+  - has optional support to use latex and dvisvgm to generate your latex
+  - function `tex2svg` added to convert Lstring to svg using dvisvgm
+- svg2luxor.jl
+  - change font scaling in svg2luxor from 1/2 to 425/1000
+  - `draw_obj` for paths can now handle "stroke" attributes
+  - `draw_obj` can handle C and S svg path instructions.
+  - warnings for missing attributes are printed only once at their first occurance
 - change default line width to 2.0 like luxor 
+
 
 ## v0.8.0 (1st of February 2022)
 - Allow Luxor v3.0